refactor out rss generation
